1.初步实现了 Map to Model

+++ src/test/java/com/wuqiushan/testModel/Courses.java
@@ -3,5 +3,5 @@ package com.wuqiushan.testModel;
 public class Courses {
 
     private String name;
-    private int    duration;
+    private double    duration;
 }
